requirements.txt
# SIT RAG Chatbot - Dependencies
# Production-ready RAG chatbot for Siddaganga Institute of Technology

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM & Embeddings (Google Gemini)
google-generativeai==0.8.0

# Vector Store
faiss-cpu==1.7.4

# Document Processing
pypdf2==3.0.1
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
numpy==1.26.3

# Testing
pytest==7.4.4
httpx==0.26.0
